Lab 1.3 incomplete (lectures complete, other in progress)

+++ students/Y2332/laboratory_works/Petrosyan_Alexandr/laboratiry_work/aud_dist/aud_dist_app/management/commands/cleardb.py
@@ -6,7 +6,8 @@ class Command(BaseCommand):
     help = "Clear all data from audience distribution project database"
 
     def handle(self, *args, **options):
-        Discipline.objects.all().delete()
-        Lecturer.objects.all().delete()
-        Group.objects.all().delete()
-        Audience.objects.all().delete()
+        if input("Are you sure? (Y/N): ") in ('y', 'Y'):
+            Discipline.objects.all().delete()
+            Lecturer.objects.all().delete()
+            Group.objects.all().delete()
+            Audience.objects.all().delete()
